txn: move gc to action module (#8953)
