+++ src/storage/txn/actions/mod.rs
@@ -8,11 +8,11 @@
 pub(crate) mod shared;
 
 pub mod acquire_pessimistic_lock;
+pub mod check_data_constraint;
 pub mod check_txn_status;
 pub mod cleanup;
 pub mod commit;
+pub mod gc;
 pub mod pessimistic_prewrite;
 pub mod prewrite;
-
-pub mod check_data_constraint;
 pub mod tests;
